+++ src/handshake.rs
@@ -8,8 +8,9 @@ mod server_hello;
 pub use certificate::certificate_from_der;
 pub(crate) use certificate::{Certificate, CertificateEntry, CertificateVerify};
 pub(crate) use client_hello::{ClientHello, ClientHelloBuilder};
+pub(crate) use extension::KeyShareEntry;
 pub(crate) use finished::finished_with_hs_hdr;
-pub(crate) use named_group::NamedGroup;
+pub use named_group::NamedGroup;
 pub(crate) use server_hello::{ServerHello, ServerHelloBuilder};
 
 use crate::alert::AlertDescription;