+++ src/handshake/extension/key_share.rs
@@ -1,5 +1,12 @@
 use crate::{alert::AlertDescription, handshake::named_group::NamedGroup, parse};
 
+/// Unrecognized KeyShare entry.
+#[derive(Debug, Clone)]
+pub(crate) struct UnrecognizedKeyShareEntry {
+    pub group: Result<NamedGroup, u16>,
+    pub key_exchange: Vec<u8>,
+}
+
 /// KeyShare entry.
 ///
 /// # References
@@ -12,36 +19,49 @@ use crate::{alert::AlertDescription, handshake::named_group::NamedGroup, parse};
 ///     opaque key_exchange<1..2^16-1>;
 /// } KeyShareEntry;
 /// ```
-#[derive(Debug)]
-pub(crate) struct KeyShareEntry {
-    pub group: Result<NamedGroup, u16>,
-    pub key_exchange: Vec<u8>,
+#[derive(Debug, Clone)]
+pub(crate) enum KeyShareEntry {
+    Secp256r1(p256::PublicKey),
+    X25519(x25519_dalek::PublicKey),
+    Unrecognized(UnrecognizedKeyShareEntry),
 }
 
 impl KeyShareEntry {
-    pub fn new_secp256r1(data: &[u8; 65]) -> Self {
-        Self {
-            group: Ok(NamedGroup::secp256r1),
-            key_exchange: data.to_vec(),
+    pub fn named_group(&self) -> Result<NamedGroup, u16> {
+        match self {
+            KeyShareEntry::Secp256r1(_) => Ok(NamedGroup::secp256r1),
+            KeyShareEntry::X25519(_) => Ok(NamedGroup::x25519),
+            KeyShareEntry::Unrecognized(ur) => ur.group,
         }
     }
 
     pub fn ser(&self) -> Vec<u8> {
-        let mut buf: Vec<u8> = Vec::with_capacity(self.key_exchange.len().saturating_add(4));
-
-        match self.group {
-            Ok(o) => buf.extend_from_slice(&o.to_be_bytes()),
-            Err(e) => buf.extend_from_slice(&e.to_be_bytes()),
+        let named_group: [u8; 2] = match self {
+            KeyShareEntry::Secp256r1(_) => NamedGroup::secp256r1.to_be_bytes(),
+            KeyShareEntry::X25519(_) => NamedGroup::x25519.to_be_bytes(),
+            KeyShareEntry::Unrecognized(unrecognized_key_share_entry) => {
+                match unrecognized_key_share_entry.group {
+                    Ok(ng) => ng.to_be_bytes(),
+                    Err(val) => val.to_be_bytes(),
+                }
+            }
         };
 
+        let key_exchange: &[u8] = match self {
+            KeyShareEntry::Secp256r1(public_key) => &public_key.to_sec1_bytes(),
+            KeyShareEntry::X25519(public_key) => public_key.as_bytes(),
+            KeyShareEntry::Unrecognized(unrecognized) => &unrecognized.key_exchange,
+        };
+
+        let mut buf: Vec<u8> = Vec::with_capacity(key_exchange.len().saturating_add(4));
+        buf.extend_from_slice(&named_group);
         buf.extend_from_slice(
-            &u16::try_from(self.key_exchange.len())
+            &u16::try_from(key_exchange.len())
                 // unwrap will not occur, length validated in constructor
                 .unwrap()
                 .to_be_bytes(),
         );
-
-        buf.extend_from_slice(&self.key_exchange);
+        buf.extend_from_slice(key_exchange);
 
         buf
     }
@@ -50,13 +70,52 @@ impl KeyShareEntry {
         let (b, group) = parse::u16("KeyShareEntry group", b)?;
         let (b, key_exchange) = parse::vec16("KeyShareEntry key_exchange", b, 1, 1)?;
 
-        Ok((
-            b,
-            Self {
-                group: NamedGroup::try_from(group),
-                key_exchange: key_exchange.to_vec(),
-            },
-        ))
+        let named_group = NamedGroup::try_from(group);
+
+        match named_group {
+            Ok(NamedGroup::secp256r1) => {
+                const P256_CLIENT_SHARE_LEN_EXPECTED: usize = 65;
+
+                if key_exchange.len() != P256_CLIENT_SHARE_LEN_EXPECTED {
+                    log::error!(
+                        "ClientHello KeyShareEntry secp256r1 has key_exchange length {} expected {}",
+                        key_exchange.len(),
+                        P256_CLIENT_SHARE_LEN_EXPECTED
+                    );
+                    return Err(AlertDescription::DecodeError)?;
+                }
+
+                match p256::PublicKey::from_sec1_bytes(key_exchange) {
+                    Ok(pk) => Ok((b, Self::Secp256r1(pk))),
+                    Err(_) => {
+                        log::error!(
+                            "ClientHello KeyShareEntry secp256r1 key_exchange data is not a valid SEC1 public key"
+                        );
+                        Err(AlertDescription::DecodeError)
+                    }
+                }
+            }
+            Ok(NamedGroup::x25519) => {
+                let key_exact: [u8; 32] = match key_exchange.try_into() {
+                    Ok(key) => key,
+                    Err(_) => {
+                        log::error!(
+                            "ClientHello KeyShareEntry x25519 has key_exchange length {} expected 32",
+                            key_exchange.len(),
+                        );
+                        return Err(AlertDescription::DecodeError)?;
+                    }
+                };
+                Ok((b, Self::X25519(x25519_dalek::PublicKey::from(key_exact))))
+            }
+            Err(_) | Ok(_) => Ok((
+                b,
+                Self::Unrecognized(UnrecognizedKeyShareEntry {
+                    group: named_group,
+                    key_exchange: key_exchange.to_vec(),
+                }),
+            )),
+        }
     }
 }
 
@@ -72,79 +131,41 @@ impl KeyShareEntry {
 /// } KeyShareClientHello;
 /// ```
 #[derive(Debug)]
-pub struct KeyShareClientHello {
-    pub secp256r1: Option<p256::PublicKey>,
-    empty: bool,
+pub(crate) struct KeyShareClientHello {
+    pub(crate) client_shares: Vec<KeyShareEntry>,
 }
 
 impl KeyShareClientHello {
-    const CLIENT_SHARE_LEN_EXPECTED: usize = 65;
-
-    pub fn deser_secp256r1(b: &[u8]) -> Result<Self, AlertDescription> {
+    pub fn deser(b: &[u8]) -> Result<Self, AlertDescription> {
         let (_, mut b) = parse::vec16("KeyShareClientHello client_shares", b, 0, 1)?;
 
-        let empty: bool = b.is_empty();
-
-        let mut secp256r1: Option<p256::PublicKey> = None;
+        let mut client_shares: Vec<KeyShareEntry> = Vec::new();
 
         while !b.is_empty() {
             let (new_b, client_share) = KeyShareEntry::deser(b)?;
             b = new_b;
 
-            if client_share.group == Ok(NamedGroup::secp256r1) {
-                let client_share_len: usize = client_share.key_exchange.len();
-                if client_share_len != Self::CLIENT_SHARE_LEN_EXPECTED {
-                    log::error!(
-                        "ClientHello KeyShareEntry secp256r1 has key_exchange length {} expected {}",
-                        client_share_len,
-                        Self::CLIENT_SHARE_LEN_EXPECTED
-                    );
-                    return Err(AlertDescription::DecodeError)?;
-                }
-
-                secp256r1 = match p256::PublicKey::from_sec1_bytes(&client_share.key_exchange) {
-                    Ok(pk) => Some(pk),
-                    Err(_) => {
-                        log::error!(
-                            "ClientHello KeyShareEntry secp256r1 key_exchange data is not a valid SEC1 public key"
-                        );
-                        return Err(AlertDescription::DecodeError)?;
-                    }
-                };
-            } else {
-                log::warn!("Unused group type {:?}", client_share.group);
-            }
+            client_shares.push(client_share);
         }
 
-        Ok(Self { secp256r1, empty })
+        Ok(Self { client_shares })
     }
 
     pub fn is_empty(&self) -> bool {
-        self.empty
+        self.client_shares.is_empty()
     }
 
-    pub fn ser_secp256r1(key: &[u8; 65]) -> Vec<u8> {
+    pub fn ser(client_shares: Vec<KeyShareEntry>) -> Vec<u8> {
         let mut ret: Vec<u8> = Vec::new();
 
-        ret.extend_from_slice(
-            u16::try_from(Self::CLIENT_SHARE_LEN_EXPECTED)
-                .unwrap()
-                .checked_add(4)
-                .unwrap()
-                .to_be_bytes()
-                .as_ref(),
-        );
+        let entries: Vec<Vec<u8>> = client_shares.iter().map(|kse| kse.ser()).collect();
+        let entries_len: usize = entries
+            .iter()
+            .fold(0, |acc, kse| acc.checked_add(kse.len()).unwrap());
 
-        ret.extend_from_slice(NamedGroup::secp256r1.to_be_bytes().as_ref());
+        ret.extend_from_slice(u16::try_from(entries_len).unwrap().to_be_bytes().as_ref());
 
-        ret.extend_from_slice(
-            u16::try_from(Self::CLIENT_SHARE_LEN_EXPECTED)
-                .unwrap()
-                .to_be_bytes()
-                .as_ref(),
-        );
-
-        ret.extend_from_slice(key);
+        entries.iter().for_each(|kse| ret.extend_from_slice(kse));
 
         ret
     }