+++ src/lib.rs
@@ -16,6 +16,7 @@ mod x509;
 pub use alert::{Alert, AlertDescription, AlertLevel};
 pub use client::{TlsClientBuilder, TlsClientStream};
 pub use error::TlsError;
+pub use handshake::NamedGroup;
 pub use server::{ServerCertificates, TlsServerBuilder, TlsServerStream};
 
 pub(crate) const GCM_TAG_LEN: usize = 16;